backend/requirements.txt
asgiref==3.5.2
certifi==2021.5.30
charset-normalizer==2.0.4
Django==2.2.16
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.4.1
drf-base64==2.0
flake8==5.0.4
Pillow==9.3.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
recipes==0.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.20.0
